Python/requirements.txt
# 核心依赖
streamlit>=1.28.0
watchdog>=3.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 进度显示
tqdm>=4.66.0

# 图像处理
opencv-python>=4.8.0

# 工具
pytz>=2023.3
typing-extensions>=4.7.0